Refactor yts_api
